Creation of propensity fitting function

Creates an propensity function that optionally utilizes cross-fitting

create.propensity.function(crossfit = TRUE, nfolds.crossfit = 10,
  cv.glmnet.args = NULL)

Arguments

crossfit
A logical value indicating whether to use cross-fitting (TRUE) or not (FALSE). Cross-fitting is more computationally intensive, but helps to prevent overfitting, see Chernozhukov, et al. (2018)

nfolds.crossfit
An integer specifying the number of folds to use for cross-fitting. Must be greater than 1

cv.glmnet.args
A list of NAMED arguments to pass to the cv.glmnet function. For example, cv.glmnet.args = list(type.measure = "mse", nfolds = 10). See cv.glmnet and glmnet for all possible options.

Value

A function which can be passed to the augment.func argument of the fit.subgroup function.

References

Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E., Hansen, C., Newey, W., & Robins, J. (2018). Double/debiased machine learning for treatment and structural parameters https://arxiv.org/abs/1608.00060

See also

Examples

library(personalized)

set.seed(123)
n.obs  <- 500
n.vars <- 15
x <- matrix(rnorm(n.obs * n.vars, sd = 3), n.obs, n.vars)

# simulate non-randomized treatment
xbetat   <- 0.5 + 0.5 * x[,7] - 0.5 * x[,9]
trt.prob <- exp(xbetat) / (1 + exp(xbetat))
trt01    <- rbinom(n.obs, 1, prob = trt.prob)

trt      <- 2 * trt01 - 1

# simulate response
# delta below drives treatment effect heterogeneity
delta <- 2 * (0.5 + x[,2] - x[,3] - x[,11] + x[,1] * x[,12] )
xbeta <- x[,1] + x[,11] - 2 * x[,12]^2 + x[,13] + 0.5 * x[,15] ^ 2
xbeta <- xbeta + delta * trt

# continuous outcomes
y <- drop(xbeta) + rnorm(n.obs, sd = 2)

aug.func <- create.augmentation.function(family = "gaussian",
                                         crossfit = TRUE,
                                         nfolds.crossfit = 10,
                                         cv.glmnet.args = list(type.measure = "mae",
                                                               nfolds = 5))

prop.func <- create.propensity.function(crossfit = TRUE,
                                        nfolds.crossfit = 10,
                                        cv.glmnet.args = list(type.measure = "mae",
                                                              nfolds = 5))

subgrp.model <- fit.subgroup(x = x, y = y,
                             trt = trt01,
                             propensity.func = prop.func,
                             augment.func = aug.func,
                             loss   = "sq_loss_lasso",
                             nfolds = 10)    # option for cv.glmnet (for ITR estimation)

summary(subgrp.model)

#> family:    gaussian 
#> loss:      sq_loss_lasso 
#> method:    weighting 
#> cutpoint:  0 
#> augmentation 
#> function: augment.func 
#> propensity 
#> function:  propensity.func 
#> 
#> benefit score: f(x), 
#> Trt recom = 1*I(f(x)>c)+0*I(f(x)<=c) where c is 'cutpoint'
#> 
#> Average Outcomes:
#>                 Recommended 0      Recommended 1
#> Received 0   -7.2029 (n = 84) -28.7306 (n = 117)
#> Received 1 -26.6205 (n = 114)  -2.9403 (n = 185)
#> 
#> Treatment effects conditional on subgroups:
#> Est of E[Y|T=0,Recom=0]-E[Y|T=/=0,Recom=0] 
#>                          19.4176 (n = 198) 
#> Est of E[Y|T=1,Recom=1]-E[Y|T=/=1,Recom=1] 
#>                          25.7903 (n = 302) 
#> 
#> NOTE: The above average outcomes are biased estimates of
#>       the expected outcomes conditional on subgroups. 
#>       Use 'validate.subgroup()' to obtain unbiased estimates.
#> 
#> ---------------------------------------------------
#> 
#> Benefit score quantiles (f(X) for 1 vs 0): 
#>      0%     25%     50%     75%    100% 
#> -35.936  -5.774   2.999  12.148  43.013 
#> 
#> ---------------------------------------------------
#> 
#> Summary of individual treatment effects: 
#> E[Y|T=1, X] - E[Y|T=0, X]
#> 
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#> -71.871 -11.547   5.997   7.096  24.295  86.025 
#> 
#> ---------------------------------------------------
#> 
#> 9 out of 15 interactions selected in total by the lasso (cross validation criterion).
#> 
#> The first estimate is the treatment main effect, which is always selected. 
#> Any other variables selected represent treatment-covariate interactions.
#> 
#>            Trt1     V2      V3     V6     V8      V9    V10     V11    V13
#> Estimate 3.7311 2.3497 -2.6714 0.7016 1.2338 -0.3431 0.9364 -1.0806 1.5617
#>             V15
#> Estimate 1.1872