Calculation of covariate-conditional treatment effects

Calculates covariate conditional treatment effects using estimated benefit scores

treatment.effects(x, ...)

# S3 method for default
treatment.effects(x, ...)

treat.effects(benefit.scores, loss = all_losses,
  method = c("weighting", "a_learning"), pi.x = NULL, ...)

# S3 method for subgroup_fitted
treatment.effects(x, ...)

Arguments

x
a fitted object from fit.subgroup()

...
not used

benefit.scores
vector of estimated benefit scores

loss
loss choice USED TO CALCULATE benefit.scores of both the M function from Chen, et al (2017) and potentially the penalty used for variable selection. See fit.subgroup for more details.

method
method choice USED TO CALCULATE benefit.scores. Either the "weighting" method or "a_learning" method. See fit.subgroup for more details

pi.x
The propensity score for each observation

Value

A List with elements delta (if the treatment effects are a difference/contrast, i.e. \(E[Y|T=1, X] - E[Y|T=-1, X]\)) and gamma (if the treatment effects are a ratio, i.e. \(E[Y|T=1, X] / E[Y|T=-1, X]\))

See also

Examples

library(personalized)

set.seed(123)
n.obs  <- 1000
n.vars <- 50
x <- matrix(rnorm(n.obs * n.vars, sd = 3), n.obs, n.vars)

# simulate non-randomized treatment
xbetat   <- 0.5 + 0.5 * x[,21] - 0.5 * x[,41]
trt.prob <- exp(xbetat) / (1 + exp(xbetat))
trt01    <- rbinom(n.obs, 1, prob = trt.prob)

trt      <- 2 * trt01 - 1

# simulate response
delta <- 2 * (0.5 + x[,2] - x[,3] - x[,11] + x[,1] * x[,12])
xbeta <- x[,1] + x[,11] - 2 * x[,12]^2 + x[,13]
xbeta <- xbeta + delta * trt

# continuous outcomes
y <- drop(xbeta) + rnorm(n.obs, sd = 2)

# time-to-event outcomes
surv.time <- exp(-20 - xbeta + rnorm(n.obs, sd = 1))
cens.time <- exp(rnorm(n.obs, sd = 3))
y.time.to.event  <- pmin(surv.time, cens.time)
status           <- 1 * (surv.time <= cens.time)

# create function for fitting propensity score model
prop.func <- function(x, trt)
{
    # fit propensity score model
    propens.model <- cv.glmnet(y = trt,
                               x = x, family = "binomial")
    pi.x <- predict(propens.model, s = "lambda.min",
                    newx = x, type = "response")[,1]
    pi.x
}

subgrp.model <- fit.subgroup(x = x, y = y,
                             trt = trt01,
                             propensity.func = prop.func,
                             loss   = "sq_loss_lasso",
                             nfolds = 5)    # option for cv.glmnet

trt_eff <- treatment.effects(subgrp.model)
str(trt_eff)

#> List of 2
#>  $ delta: num [1:1000] -11.55 -9.8 4.52 -7.03 -26.16 ...
#>  $ gamma: logi NA

print(summary(trt_eff$delta))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#> -57.436 -13.833  -2.261  -1.396  10.513  49.163 

library(survival)
subgrp.model.cox <- fit.subgroup(x = x, y = Surv(y.time.to.event, status),
                           trt = trt01,
                           propensity.func = prop.func,
                           loss   = "cox_loss_lasso",
                           nfolds = 5)              # option for cv.glmnet

trt_eff_c <- treatment.effects(subgrp.model.cox)
str(trt_eff_c)

#> List of 2
#>  $ delta: logi NA
#>  $ gamma: num [1:1000] 0.618 0.772 1.112 0.863 0.393 ...

print(summary(trt_eff_c$gamma))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  0.1886  0.7587  1.1040  1.2425  1.5352  4.3112