National Supported Work Study Data

The LaLonde dataset comes from the National Supported Work Study, which sought to evaluate the effectiveness of an employment trainining program on wage increases.

LaLonde

Format

A data frame with 722 observations and 12 variables:

outcome

whether earnings in 1978 are larger than in 1975; 1 for yes, 0 for no

treat

whether the individual received the treatment; "Yes" or "No"

age

age in years

educ

education in years

black

black or not; factor with levels "Yes" or "No"

hisp

hispanic or not; factor with levels "Yes" or "No"

white

white or not; factor with levels "Yes" or "No"

marr

married or not; factor with levels "Yes" or "No"

nodegr

No high school degree; factor with levels "Yes" (for no HS degree) or "No"

log.re75

log of earnings in 1975

u75

unemployed in 1975; factor with levels "Yes" or "No"

wts.extrap

extrapolation weights to the 1978 Panel Study for Income Dynamics dataset

Source

The National Supported Work Study.

References

LaLonde, R.J. 1986. "Evaluating the econometric evaulations of training programs with experimental data." American Economic Review, Vol.76, No.4, pp. 604-620.

Egami N, Ratkovic M, Imai K (2017). "FindIt: Finding Heterogeneous Treatment Effects." R package version 1.1.2, https://CRAN.R-project.org/package=FindIt.

Examples

data(LaLonde)
y <- LaLonde$outcome

trt <- LaLonde$treat

x.varnames <- c("age", "educ", "black", "hisp", "white",
                "marr", "nodegr", "log.re75", "u75")

# covariates
data.x <- LaLonde[, x.varnames]

# construct design matrix (with no intercept)
x <- model.matrix(~ -1 + ., data = data.x)

const.propens <- function(x, trt)
{
    mean.trt <- mean(trt == "Trt")
    rep(mean.trt, length(trt))
}

subgrp_fit_w <- fit.subgroup(x = x, y = y, trt = trt,
    loss = "logistic_loss_lasso",
    propensity.func = const.propens,
    cutpoint = 0,
    type.measure = "auc",
    nfolds = 10)

summary(subgrp_fit_w)

#> family:  binomial 
#> loss:    logistic_loss_lasso 
#> method:  weighting 
#> propensity 
#> function: propensity.func 
#> 
#> Average Outcomes:
#>               Recommended Ctrl  Recommended Trt
#> Received Ctrl  0.7292 (n = 48) 0.5146 (n = 377)
#> Received Trt   0.5714 (n = 28) 0.6059 (n = 269)
#> 
#> Ctrl effect among recommended Ctrl   Trt effect among recommended Trt 
#>                    0.1577 (n = 76)                   0.0914 (n = 646) 
#> 
#> Benefit score quantiles: 
#>      0%     25%     50%     75%    100% 
#> -0.1811  0.1341  0.1341  0.1341  0.2980 
#> 
#> 2 out of 10 variables selected in total by the lasso (cross validation criterion).
#> 
#>         Estimate
#> Trt       0.1341
#> hispYes  -0.3152
#> marrYes   0.1639