Fitting subgroup identification models

Fits subgroup identification model class of Chen, et al (2017)

fit.subgroup(x, y, trt, propensity.func, loss = c("sq_loss_lasso",
  "logistic_loss_lasso", "cox_loss_lasso", "sq_loss_lasso_gam",
  "logistic_loss_lasso_gam", "sq_loss_gam", "logistic_loss_gam", "sq_loss_gbm",
  "abs_loss_gbm", "logistic_loss_gbm", "cox_loss_gbm"),
  method = c("weighting", "a_learning"), augment.func = NULL,
  cutpoint = 0, larger.outcome.better = TRUE, retcall = TRUE, ...)

Arguments

x
The design matrix (not including intercept term)

y
The response vector

trt
treatment vector with each element equal to a 0 or a 1, with 1 indicating treatment status is active.

propensity.func
function that inputs the design matrix x and the treatment vector trt and outputs the propensity score, ie Pr(trt = 1 | X = x). Function should take two arguments 1) x and 2) trt. See example below. For a randomized controlled trial this can simply be a function that returns a constant equal to the proportion of patients assigned to the treatment group, i.e.: propensity.func = function(x, trt) 0.5.

loss
choice of both the M function from Chen, et al (2017) and potentially the penalty used for variable selection. All loss options starting with sq_loss use M(y, v) = (v - y) ^ 2, all options starting with logistic_loss use the logistic loss: M(y, v) = y * log(1 + exp-v), and all options starting with cox_loss use the negative partial likelihood loss for the Cox PH model. All options ending with lasso have a lasso penalty added to the loss for variable selection. sq_loss_lasso_gam and logistic_loss_lasso_gam first use the lasso to select variables and then fit a generalized additive model with nonparametric additive terms for each selected variable. sq_loss_gam involves a squared error loss with a generalized additive model and no variable selection. sq_loss_gbm involves a squared error loss with a gradient-boosted decision trees model for the benefit score; this allows for flexible estimation using machine learning and can be useful when the underlying treatment-covariate interaction is complex.

"sq_loss_lasso" - M(y, v) = (v - y) ^ 2 with linear model and lasso penalty

"logistic_loss_lasso" - M(y, v) = -[yv - log(1 + exp{-v})] with with linear model and lasso penalty

"cox_loss_lasso" - M corresponds to the negative partial likelihood of the cox model with linear model and additionally a lasso penalty

"sq_loss_lasso_gam" - M(y, v) = (v - y) ^ 2 with variables selected by lasso penalty and generalized additive model fit on the selected variables

"logistic_loss_lasso_gam" - M(y, v) = y * log(1 + exp-v) with variables selected by lasso penalty and generalized additive model fit on the selected variables

"sq_loss_gam" - M(y, v) = (v - y) ^ 2 with generalized additive model fit on all variables

"logistic_loss_gam" - M(y, v) = y * log(1 + exp-v) with generalized additive model fit on all variables

"sq_loss_gbm" - M(y, v) = (v - y) ^ 2 with gradient-boosted decision trees model

"abs_loss_gbm" - M(y, v) = |v - y| with gradient-boosted decision trees model

"logistic_loss_gbm" - M(y, v) = -[yv - log(1 + exp{-v})] with gradient-boosted decision trees model

"cox_loss_gbm" - M corresponds to the negative partial likelihood of the cox model with gradient-boosted decision trees model

method
subgroup ID model type. Either the weighting or A-learning method of Chen et al, (2017)

augment.func
function which inputs the response y, the covariates x, and trt and outputs predicted values for the response using a model constructed with x. augment.func() can also be simply a function of x and y. This function is used for efficiency augmentation. When the form of the augmentation function is correct, it can provide efficient estimation of the subgroups Example 1: augment.func <- function(x, y) {lmod <- lm(y ~ x); return(fitted(lmod))}

Example 2: augment.func <- function(x, y, trt) {lmod <- lm(y ~ x + trt); return(fitted(lmod))}

cutpoint
numeric value for patients with benefit scores above which (or below which if larger.outcome.better = FALSE) will be recommended to be in the treatment group

larger.outcome.better
boolean value of whether a larger outcome is better/preferable. Set to TRUE if a larger outcome is better/preferable and set to FALSE if a smaller outcome is better/preferable. Defaults to TRUE.

retcall
boolean value. if TRUE then the passed arguments will be saved. Do not set to FALSE if the validate.subgroupu() function will later be used for your fitted subgroup model. Only set to FALSE if memory is limited as setting to codeTRUE saves the design matrix to the fitted object

...
options to be passed to underlying fitting function. For all loss options with lasso, this will be passed to cv.glmnet and for all loss options with mcp this will be passed to cv.ncvreg. Note that for all loss options that use gam() from the mgcv package, the user cannot supply the gam argument method because it is also an argument of fit.subgroup, so instead, to change the gam method argument, instead supply method.gam, ie method.gam = "REML".

References

Chen, S., Tian, L., Cai, T. and Yu, M. (2017), A general statistical framework for subgroup identification and comparative treatment scoring. Biometrics. doi:10.1111/biom.12676 http://onlinelibrary.wiley.com/doi/10.1111/biom.12676/abstract

See also

validate.subgroup for function which creates validation results for subgroup identification models, predict.subgroup_fitted for a prediction function for fitted models from fit.subgroup, and plot.subgroup_fitted for a function which plots results from fitted models from fit.subgroup.

Examples

library(personalized)

set.seed(123)
n.obs  <- 1000
n.vars <- 50
x <- matrix(rnorm(n.obs * n.vars, sd = 3), n.obs, n.vars)

# simulate non-randomized treatment
xbetat   <- 0.5 + 0.5 * x[,21] - 0.5 * x[,41]
trt.prob <- exp(xbetat) / (1 + exp(xbetat))
trt01    <- rbinom(n.obs, 1, prob = trt.prob)

trt      <- 2 * trt01 - 1

# simulate response
delta <- 2 * (0.5 + x[,2] - x[,3] - x[,11] + x[,1] * x[,12] )
xbeta <- x[,1] + x[,11] - 2 * x[,12]^2 + x[,13] + 0.5 * x[,15] ^ 2
xbeta <- xbeta + delta * trt

# continuous outcomes
y <- drop(xbeta) + rnorm(n.obs, sd = 2)

# binary outcomes
y.binary <- 1 * (xbeta + rnorm(n.obs, sd = 2) > 0 )

# time-to-event outcomes
surv.time <- exp(-20 - xbeta + rnorm(n.obs, sd = 1))
cens.time <- exp(rnorm(n.obs, sd = 3))
y.time.to.event  <- pmin(surv.time, cens.time)
status           <- 1 * (surv.time <= cens.time)

# create function for fitting propensity score model
prop.func <- function(x, trt)
{
    # fit propensity score model
    propens.model <- cv.glmnet(y = trt,
                               x = x, family = "binomial")
    pi.x <- predict(propens.model, s = "lambda.min",
                    newx = x, type = "response")[,1]
    pi.x
}

subgrp.model <- fit.subgroup(x = x, y = y,
                           trt = trt01,
                           propensity.func = prop.func,
                           loss   = "sq_loss_lasso",
                           nfolds = 5)              # option for cv.glmnet

summary(subgrp.model)

#> family:  gaussian 
#> loss:    sq_loss_lasso 
#> method:  weighting 
#> 
#> Average Outcomes:
#>                  Recommended Trt  Recommended Ctrl
#> Received Trt   -2.7736 (n = 256) -22.144 (n = 321)
#> Received Ctrl -21.3209 (n = 199) -6.6962 (n = 224)
#> 
#>  Trt  Effect Among Recommended Trt Ctrl Effect Among Recommended Ctrl 
#>                  18.5473 (n = 455)                  15.4479 (n = 545) 
#> 
#> Benefit score quantiles: 
#>      0%     25%     50%     75%    100% 
#> -26.984  -6.531  -1.056   4.878  22.578 
#> 
#> 12 variables selected by the lasso (cross validation criterion).
#> 
#>     Estimate
#> Trt  -0.8824
#> V1    0.2072
#> V2    1.8824
#> V3   -1.9520
#> V11  -0.4108
#> V12  -0.1054
#> V13   0.5645
#> V21  -0.1077
#> V22  -0.1024
#> V41   0.0660
#> V42  -0.2235
#> V47  -0.0421

# fit lasso + gam model with REML option for gam

subgrp.modelg <- fit.subgroup(x = x, y = y,
                            trt = trt01,
                            propensity.func = prop.func,
                            loss   = "sq_loss_lasso_gam",
                            method.gam = "REML",     # option for gam
                            nfolds = 5)              # option for cv.glmnet

subgrp.modelg

#> family:  gaussian 
#> loss:    sq_loss_lasso_gam 
#> method:  weighting 
#> 
#> Average Outcomes:
#>                 Recommended Trt   Recommended Ctrl
#> Received Trt   2.1198 (n = 311) -31.8704 (n = 266)
#> Received Ctrl -8.0685 (n = 238) -20.6621 (n = 185)
#> 
#>  Trt  Effect Among Recommended Trt Ctrl Effect Among Recommended Ctrl 
#>                  10.1883 (n = 549)                  11.2082 (n = 451) 
#> 
#> Benefit score quantiles: 
#>       0%      25%      50%      75%     100% 
#> -143.120  -12.770    2.529   14.880   45.175 

subgrp.model.bin <- fit.subgroup(x = x, y = y.binary,
                           trt = trt01,
                           propensity.func = prop.func,
                           loss   = "logistic_loss_lasso",
                           type.measure = "auc",    # option for cv.glmnet
                           nfolds = 5)              # option for cv.glmnet

subgrp.model.bin

#> family:  binomial 
#> loss:    logistic_loss_lasso 
#> method:  weighting 
#> 
#> Average Outcomes:
#>                Recommended Trt Recommended Ctrl
#> Received Trt  0.5309 (n = 324) 0.1502 (n = 253)
#> Received Ctrl 0.2292 (n = 240) 0.4863 (n = 183)
#> 
#>  Trt  Effect Among Recommended Trt Ctrl Effect Among Recommended Ctrl 
#>                   0.3017 (n = 564)                   0.3361 (n = 436) 
#> 
#> Benefit score quantiles: 
#>       0%      25%      50%      75%     100% 
#> -1.49663 -0.24413  0.07284  0.43341  1.49749 

library(survival)
subgrp.model.cox <- fit.subgroup(x = x, y = Surv(y.time.to.event, status),
                           trt = trt01,
                           propensity.func = prop.func,
                           loss   = "cox_loss_lasso",
                           nfolds = 5)              # option for cv.glmnet

subgrp.model.cox

#> family:  cox 
#> loss:    cox_loss_lasso 
#> method:  weighting 
#> 
#> Average Outcomes:
#>                 Recommended Trt   Recommended Ctrl
#> Received Trt  68.8973 (n = 265) 152.5531 (n = 312)
#> Received Ctrl 19.9534 (n = 194) 193.8419 (n = 229)
#> 
#>  Trt  Effect Among Recommended Trt Ctrl Effect Among Recommended Ctrl 
#>                  48.9438 (n = 459)                  41.2888 (n = 541) 
#> 
#> Benefit score quantiles: 
#>       0%      25%      50%      75%     100% 
#> -1.26474 -0.38744 -0.05997  0.27320  1.47516