Printing results for fitted subgroup identification models

Prints results for estimated subgroup treatment effects

Prints summary results for estimated subgroup treatment effects

# S3 method for subgroup_fitted
print(x, digits = max(getOption("digits") - 3, 3), ...)

# S3 method for subgroup_validated
print(
  x,
  digits = max(getOption("digits") - 3, 3),
  sample.pct = FALSE,
  which.quant = NULL,
  ...
)

# S3 method for subgroup_summary
print(x, p.value = 0.001, digits = max(getOption("digits") - 3, 3), ...)

Arguments

x

a fitted object from either fit.subgroup, validate.subgroup, or summarize.subgroups()

digits

minimal number of significant digits to print.

...

further arguments passed to or from print.default.

sample.pct

boolean variable of whether to print the percent of the test sample within each subgroup. If false the sample size itself, not the percent is printed. This may not be informative if the test sample size is much different from the total sample size

which.quant

when validate.subgroup() is called with a vector of quantile values specified for benefit.score.quantiles, i.e. benefit.score.quantiles = c(0.25, 0.5, 0.75), the argument which.quant can be a vector of indexes specifying which quantile cutoff value validation results to display, i.e. which.quant = c(1,3) in the above example results in the display of validation results for subgroups defined by cutoff values of the benefit score defined by the 25th abnd 75th quantiles of the benefit score

p.value

a p-value threshold for mean differences below which covariates will be displayed. P-values are adjusted for multiple comparisons by the Hommel approach. For example, setting p.value = 0.05 will display all covariates that have a significant difference between subgroups with p-value less than 0.05. Defaults to 0.001.

See also