Validating fitted subgroup identification models

Validates subgroup treatment effects for fitted subgroup identification model class of Chen, et al (2017)

validate.subgroup(model, B = 50L, method = c("training_test_replication",
  "boot_bias_correction"), train.fraction = 0.5, parallel = FALSE)

Arguments

model
fitted model object returned by fit.subgroup() function

B
integer. number of bootstrap replications or refitting replications.

method
validation method. "boot_bias_correction" for the bootstrap bias correction method of Harrell, et al (1996) or "training_test_replication" for repeated training and test splitting of the data (train.fraction should be specified for this option)

train.fraction
fraction (between 0 and 1) of samples to be used for training in training/test replication. Only used for method = "training_test_replication"

parallel
Should the loop over replications be parallelized? If FALSE, then no, if TRUE, then yes. If user sets parallel = TRUE and the fitted fit.subgroup() object uses the parallel version of an internal model, say for cv.glmnet(), then the internal parallelization will be overridden so as not to create a conflict of parallelism.

References

Chen, S., Tian, L., Cai, T. and Yu, M. (2017), A general statistical framework for subgroup identification and comparative treatment scoring. Biometrics. doi:10.1111/biom.12676

Harrell, F. E., Lee, K. L., and Mark, D. B. (1996). Tutorial in biostatistics multivariable prognostic models: issues in developing models, evaluating assumptions and adequacy, and measuring and reducing errors. Statistics in medicine, 15, 361-387. doi:10.1002/(SICI)1097-0258(19960229)15:4<361::AID-SIM168>3.0.CO;2-4

See also

fit.subgroup for function which fits subgroup identification models, plot.subgroup_validated for plotting of validation results, and print.subgroup_validated for arguments for printing options for validate.subgroup().

Examples

library(personalized)

set.seed(123)
n.obs  <- 500
n.vars <- 20
x <- matrix(rnorm(n.obs * n.vars, sd = 3), n.obs, n.vars)

# simulate non-randomized treatment
xbetat   <- 0.5 + 0.5 * x[,11] - 0.5 * x[,13]
trt.prob <- exp(xbetat) / (1 + exp(xbetat))
trt01    <- rbinom(n.obs, 1, prob = trt.prob)

trt      <- 2 * trt01 - 1

# simulate response
delta <- 2 * (0.5 + x[,2] - x[,3] - x[,11] + x[,1] * x[,12])
xbeta <- x[,1] + x[,11] - 2 * x[,12]^2 + x[,13]
xbeta <- xbeta + delta * trt

# continuous outcomes
y <- drop(xbeta) + rnorm(n.obs, sd = 2)

# create function for fitting propensity score model
prop.func <- function(x, trt)
{
    # fit propensity score model
    propens.model <- cv.glmnet(y = trt,
                               x = x, family = "binomial")
    pi.x <- predict(propens.model, s = "lambda.min",
                    newx = x, type = "response")[,1]
    pi.x
}

subgrp.model <- fit.subgroup(x = x, y = y,
                             trt = trt01,
                             propensity.func = prop.func,
                             loss   = "sq_loss_lasso",
                             nfolds = 5)    # option for cv.glmnet

subgrp.model$subgroup.trt.effects

#> $subgroup.effects
#> 0 effect among recommended 0 1 effect among recommended 1 
#>                     13.84725                     13.26465 
#> 
#> $avg.outcomes
#>            Recommended 0 Recommended 1
#> Received 0     -9.827756     -27.66231
#> Received 1    -23.675004     -14.39765
#> 
#> $sample.sizes
#>            Recommended 0 Recommended 1
#> Received 0            50           170
#> Received 1           159           121
#> 
#> $overall.subgroup.effect
#> [1] 12.67388
#> 

x.test <- matrix(rnorm(10 * n.obs * n.vars, sd = 3), 10 * n.obs, n.vars)

# simulate non-randomized treatment
xbetat.test   <- 0.5 + 0.5 * x.test[,11] - 0.5 * x.test[,13]
trt.prob.test <- exp(xbetat.test) / (1 + exp(xbetat.test))
trt01.test    <- rbinom(10 * n.obs, 1, prob = trt.prob.test)

trt.test      <- 2 * trt01.test - 1

# simulate response
delta.test <- 2 * (0.5 + x.test[,2] - x.test[,3] - x.test[,11] + x.test[,1] * x.test[,12])
xbeta.test <- x.test[,1] + x.test[,11] - 2 * x.test[,12]^2 + x.test[,13]
xbeta.test <- xbeta.test + delta.test * trt.test

y.test <- drop(xbeta.test) + rnorm(10 * n.obs, sd = 2)

valmod <- validate.subgroup(subgrp.model, B = 10,
                            method = "training_test",
                            train.fraction = 0.75)
valmod

#> family:  gaussian 
#> loss:    sq_loss_lasso 
#> method:  weighting 
#> 
#> validation method:  training_test_replication 
#> iterations:  10 
#> 
#> Average Test Set Outcomes:
#>                               Recommended 0                    Recommended 1
#> Received 0 -14.3177 (SE = 8.8949, n = 13.7)   -25.1564 (SE = 3.6887, n = 43)
#> Received 1   -22.7101 (SE = 4.1061, n = 37) -18.0328 (SE = 5.6593, n = 31.3)
#> 
#>    0 effect among recommended 0    1 effect among recommended 1 
#> 8.3924 (SE = 10.1312, n = 50.7)  7.1236 (SE = 7.0019, n = 74.3) 

bene.score.test <- subgrp.model$predict(x.test)

mean(y.test[bene.score.test > 0 & trt01.test == 1]) -
       mean(y.test[bene.score.test > 0 & trt01.test == 0])

#> [1] 13.18799

mean(y.test[bene.score.test <= 0 & trt01.test == 0]) -
       mean(y.test[bene.score.test <= 0 & trt01.test == 1])

#> [1] 15.84766

quantile(valmod$boot.results[[1]][,1], c(0.025, 0.975))

#>     2.5%    97.5% 
#> -5.24051 25.42524 

quantile(valmod$boot.results[[1]][,2], c(0.025, 0.975))

#>      2.5%     97.5% 
#> -4.673896 16.891682